// src/app/@core/models/printouts/report-forms/optional-report-sections.ts
export interface OptionalReportSections {
	showClassTeacherRemarks: boolean;
	showPrincipalRemarks: boolean;
	showClassTeacherSignature: boolean;
	showPrincipalSignature: boolean;
	showParentSignatureSlot: boolean;
	showOverallStudentRank: boolean;
	showStreamStudentRank: boolean;
	showGPA: boolean;
	showCredentials: boolean;
	showGradeDescriptors: boolean;
}